+++ api/generatePost.ts
@@ -1,89 +1,5 @@
 import type { VercelRequest, VercelResponse } from "@vercel/node";
-import { ChatOpenAI } from "@langchain/openai";
-import { PromptTemplate } from "@langchain/core/prompts";
-import { StringOutputParser } from "@langchain/core/output_parsers";
-
-const prompt = PromptTemplate.fromTemplate(
-  `<Instruction>
-당신은 인스타그램 매거진 계정의 소셜 미디어 매니저입니다. 주로 20-30대 트렌드에 민감한 독자층을 대상으로 합니다.
-Inputs 태그 내에 제시된 자료를 바탕으로, 흥미롭고 정보가 풍부한 인스타그램 포스트를 작성해야 합니다.
-
-기존 게시글의 스타일을 유지하기 위해 Examples 태그 내의 예시들을 참고하세요. 일관된 브랜드 이미지를 유지하는 것이 중요합니다.
-</Instruction>
-
-<Guidelines>
-- 포스트의 톤앤매너를 유지하기 위해 다음 지침을 따르세요:
-  1. 간결하고 직관적인 문장을 사용하세요.
-  2. 현재형 시제를 주로 사용하세요.
-  3. 브랜드나 제품명을 언급할 때는 항상 인스타그램 계정을 태그하세요. (예: @brandname)
-  4. 이모지를 적절히 사용하되, 과도하게 사용하지 마세요. 주로 문장 끝이나 중요한 키워드 옆에 배치하세요.
-  5. 공식적이지만 친근한 어조를 유지하세요.
-
-- 전체 글은 4-5개의 문장으로 구성하되, 총 글자 수는 350자를 넘지 않도록 해주세요.
-- 글의 주요 단어를 활용하여 해시태그를 만들어주세요. 순서는 #eyesmag #영어1 #영어2 ... #한글1 #한글2 입니다.
-- 권유문(예: 초대합니다, 경험하세요 등)은 사용하지 마세요.
-- 느낌표를 사용하지 마세요.
-- 만약 인용을 할 경우, 인용 표기 기호는 다음을 참고해서 사용해주세요.
-  - 방송/전시/영화/책 : <>
-  - 앨범명 : []
-  - 음원, 공연 : ''
-
-- 여러 개의 참조 자료가 제공된 경우, 가장 중요하고 관련성 높은 정보를 선별하여 포스트에 포함시키세요.
-</Guidelines>
-
-<Examples>
-  <Example>
-  롯데워터파크(@lottewaterpark)가 반려견과 함께할 수 있는 ’댕댕워터파크‘를 오픈합니다.🐕 매년 봄과 가을에만 만날 수 있는 ’댕댕워터파크‘는 약 1,000평 크기의 국내 최대 규모 워터파크로 자이언트 아쿠아 플렉스와 수중 놀이터 티키풀을 중심으로 운영되는데요.🌊 맹견으로 지정된 5종을 제외한 모든 견종이 입장 가능하며, 수중 20cm로 운영되어 소형견도 안심하고 이용할 수 있습니다.🐶 반려견 패션쇼와 장기자랑 대회 등 다양한 행사가 열릴 ’댕댕워터파크‘는 오는 오는 8월 29일부터 오픈하며, 9월 8일까지는 상시운영, 9월 9일부터는 주말, 공휴일, 10월 4일에만 운영됩니다. 📸 @lottewaterpark
-  </Example>
-
-  <Example>
-  유니클로(@uniqlokr)가 롯데월드몰 내 2개 층, 무려 1천 평에 달하는 국내 최대 규모의 매장을 오픈합니다.👋🏻 이곳에서는 ‘리유니클로 스튜디오’와 ‘유티미!’와 같은 새로운 서비스를 국내 최초로 경험할 수 있는데요. 먼저 ‘리유니클로 스튜디오’는 구멍, 찢어짐, 솔기, 패치워크 등을 수선할 수 있는 서비스와 자수 패턴을 활용한 의류 커스터마이징 서비스를 제공하며, 총 800여 가지의 다양한 스티커를 활용해 나만의 티셔츠를 만들 수 있는 커스터마이징 서비스 ‘유티미!’도 준비되었습니다.✔️ 공식 오픈일은 오는 9월 13일. 직접 방문해 이곳에만 있는 특별한 서비스를 경험해 보세요. 자세한 내용은 매거진 사이트에서 확인할 수 있습니다. 📸 유니클로.
-  </Example>
-  
-  <Example>
-  디올(@dior)이 앰버서더 투모로우바이투게더(@txt_bighit)와 함께한 24 겨울 남성 컬렉션 비주얼을 공개했습니다.🤍 멤버들은 기능적이면서 서정적인 디자인으로 승화된 이번 컬렉션을 착용하여 우아한 면모를 뽐냈는데요. 재해석된 디올의 대담한 볼륨과 샤프한 컷이 멤버들과 조화롭게 어우러진 모습. 투모로우바이투게더와 함께한 디올 새 컬렉션은 전국 부티크와 공식 홈페이지에서 만나볼 수 있습니다. 📸 디올
-  </Example>
-
-  <Example>
-  한화이글스(@hanwhaeagles_soori)가 두산베어스(@doosanbears.1982)와의 경기에서 3연전 전승을 기록, 스윕을 달성했습니다.⚾️ 무려 7020일, 19년만. 류현진(@hyunjinryu325)이 선발로 등판하여 맹활약 했고 팽팽한 투수전 끝에 3-1 승리를 기록했는데요. 이로써 올 시즌 상대 전적에서도 두산 상대 9승 6패를 기록, 우위를 확정 지었으며 이 역시 2011년 이후 13년 만의 기적 같은 일이라고 합니다.☄️ 앞으로도 이어질 한화의 활약을 기대해 보세요! 📸 @hanwhaeagles_soori
-  </Example>
-
-  <Example>
-  서울시(@seoul_official)가 도시 브랜드 ‘서울, 마이 소울’ 출범 1주년을 맞아 바프(@hbaf_official)와 ‘서울 아몬드’를 선보입니다.🏈 K문화 중 하나인 ‘한강에서 치킨 먹기’를 모티브로, 후라이드치킨 맛, 양념치킨 맛, 간장치킨 맛 총 3가지로 구성됐는데요.🍗 제품은 내달 10일 전국 바프 오프라인 매장에서 만나볼 수 있으며, 19일부터 바프 네이버 온라인 스토어에서 판매될 예정입니다.🛒 📸 서울시
-  </예시
-
-  <Example>
-  기름이 어떻게 자동차를 움직일까?🤔 HD현대오일뱅크(@hdhyundaioilbank.official)가 정유 업계 최초로 그 비밀을 밝힌 영상을 공개했습니다.✨ 독특한 크리에이티브로 탄탄한 팬층을 보유한 크리에이티브 에이전시 돌고래유괴단(@dolphiners_films)과 함께한 본 캠페인은 차를 움직이기 위해 고군분투하는 ‘오일전사’들의 모습을 보여주며 HD현대오일뱅크의 고객을 향한 진심을 유쾌하게 보여주는 콘셉트로 풀어냈죠.🚙 배우 유지태(@jt_db)와 김동준(@super_d.j)이 열연을 펼친 해당 영상의 풀버전은 HD현대오일뱅크 유튜브 채널 및 HD현대(@hd.hyundai) 인스타그램 채널에서 감상할 수 있습니다. 🎥 HD현대오일뱅크
-  </Example>
-</Examples>
-
-<Inputs>
-  {references}
-</Inputs>`
-);
-
-const model = new ChatOpenAI({
-  model: "gpt-4o-mini",
-  apiKey: process.env.VUE_APP_OPENAI_API_KEY, // 서버에서 환경 변수로 설정
-});
-
-const parser = new StringOutputParser();
-
-function indentText(text: string, level: number): string {
-  const indent = "\t".repeat(level);
-  return text
-    .split("\n")
-    .map((line) => indent + line)
-    .join("\n");
-}
-
-function formatReference(reference: { text: string }): string {
-  const openingTag = indentText(`<ReferenceSource>`, 1);
-  const closingTag = indentText(`</ReferenceSource>`, 1);
-  const indentedText = indentText(reference.text, 2);
-
-  return `${openingTag}\n${indentedText}\n${closingTag}`;
-}
+import { generatePost } from "../backend/services/generatePostService.js";
 
 export default async function handler(req: VercelRequest, res: VercelResponse) {
   const { references } = req.body;
@@ -92,21 +8,11 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
     return res.status(400).json({ error: "Invalid input" });
   }
 
-  const formattedReferences = references
-    .map((ref) => formatReference(ref))
-    .join("\n\n");
-
-  const input = {
-    references: formattedReferences,
-  };
-
   try {
-    const chain = prompt.pipe(model).pipe(parser);
-    const result = await chain.invoke(input);
-
+    const result = await generatePost(references);
     return res.status(200).json({ result });
   } catch (error) {
-    console.error("Error invoking the chain:", error);
+    console.error("Error generating post:", error);
     return res.status(500).json({ error: "Failed to generate post" });
   }
 }
